# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Tareas",
        "description": "Documentación de la API para gestionar tareas.",
        "contact": {
            "email": "contacto@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/tareas": {
            "get": {
                "summary": "Obtener todas las tareas",
                "operationId": "6733ab2dc67122575c368d93ff4c1308",
                "responses": {
                    "200": {
                        "description": "Lista de tareas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Tarea"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    }
                }
            },
            "post": {
                "summary": "Crear una nueva tarea",
                "operationId": "cff34cbb3cdf575de33470453733f08c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Tarea"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tarea creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Tarea"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    }
                }
            }
        },
        "/api/tareas/{tarea}": {
            "get": {
                "summary": "Obtener una tarea por ID",
                "operationId": "016a666faf1eddd91659002da02de43d",
                "parameters": [
                    {
                        "name": "tarea",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarea encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Tarea"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tarea no encontrada"
                    }
                }
            },
            "put": {
                "summary": "Actualizar una tarea por ID",
                "operationId": "30a9676bd5b8b8123b8f297be0ed782a",
                "parameters": [
                    {
                        "name": "tarea",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Tarea"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tarea actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Tarea"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tarea no encontrada"
                    }
                }
            },
            "delete": {
                "summary": "Eliminar una tarea por ID",
                "operationId": "8d10da3eddc7281ad9243f16f9cb5388",
                "parameters": [
                    {
                        "name": "tarea",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarea eliminada exitosamente"
                    },
                    "404": {
                        "description": "Tarea no encontrada"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Tarea": {
                "properties": {
                    "id": {
                        "description": "ID de la tarea",
                        "type": "integer",
                        "format": "int64"
                    },
                    "titulo": {
                        "description": "Título de la tarea",
                        "type": "string"
                    },
                    "descripcion": {
                        "description": "Descripción de la tarea",
                        "type": "string"
                    },
                    "fecha_vencimiento": {
                        "description": "Fecha de vencimiento de la tarea",
                        "type": "string",
                        "format": "date"
                    },
                    "estado": {
                        "description": "Estado de la tarea",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    }
}